requirements.txt
aiohttp
bs4
nltk
unidecode
aiodns
cchardet
discord.py
